move around project structure
interfaces are folders with associated duck types
