+++ lib/game_presenter/game_presenter.rb
@@ -1,4 +1,4 @@
-require_relative '../lib/messages'
+require_relative '../messages'
 
 class GamePresenter
     attr_accessor :presenter
@@ -21,28 +21,5 @@ class GamePresenter
 
     def game_setting_IO(message:)
         presenter.game_setting_IO(message: message)
-    end
-end
-
-class ConsoleIO
-    include Messages
-    
-    def get_input(input=nil)
-        return gets.chomp
-    end
-
-    def output_message(message)
-        puts message
-    end
-
-    def clear
-        print "\e[2J\e[f"
-    end
-
-    def game_setting_IO(message:)
-        clear
-        output_message(start_banner)
-        output_message(message)
-        return get_input
     end
 end
