+++ lib/game_presenter/test_io.rb
