--- lib/game_setting.rb
@@ -1,68 +1,0 @@
-require_relative '../lib/validator'
-require_relative '../lib/messages'
-require_relative '../lib/setting_types'
-
-class GameSetting
-    attr_accessor :setting, :input
-
-    def initialize(setting:)
-        @setting = setting
-    end
-
-    def valid?(input: nil)
-        @input = input
-        setting.valid?(input: input)
-    end
-
-    def name
-        setting.name
-    end
-
-    def message
-        setting.message
-    end
-
-    def clean(input:)
-        setting.clean(input: input)
-    end
-end
-
-class OpponentType
-    include Messages
-    include Validator
-    include SettingTypes
-    attr_accessor :name, :message
-    
-    def initialize
-        @name = opponent_type
-        @message = opponent_type_msg
-    end
-
-    def valid?(input: nil)
-        opponent_type_valid?(input: input)
-    end
-
-    def clean(input:)
-        return input
-    end
-end
-
-class BoardSize
-    include Messages
-    include Validator
-    include SettingTypes
-    attr_accessor :name, :message
-    
-    def initialize
-        @name = board_size
-        @message = board_size_msg
-    end
-
-    def valid?(input: nil)
-        board_size_valid?(input: input)
-    end
-
-    def clean(input:)
-        return input.to_i
-    end
-end
