+++ lib/game_setting/setting_types.rb
