+++ lib/player/player.rb
@@ -1,0 +1,15 @@
+require_relative '../messages'
+
+class Player
+    attr_accessor :player, :token
+
+    def initialize(player:, token:) 
+        @player = player
+        @token = token
+        player.token = token
+    end
+
+    def move(board:, presenter:)
+        player.move(board: board, presenter: presenter)
+    end
+end
