+++ lib/player/player_types.rb
@@ -1,0 +1,9 @@
+module PlayerType
+    def human
+        return "H"
+    end
+
+    def computer
+        return "C"
+    end
+end
