+++ spec/validator_spec.rb
@@ -1,6 +1,6 @@
 require_relative '../lib/validator'
 require_relative '../lib/board'
-require_relative '../lib/player'
+require_relative '../lib/player/player_types'
 require_relative '../lib/token'
 
 describe 'Validator' do
